+++ data-raw/compile_tidywater_data.R
@@ -39,6 +39,10 @@ mweights <- data.frame(
   f = 18.9984,
   mn = 54.938,
   nh4 = 18.04,
+  nh3 = 17.031,
+  nh2cl = 51.48,
+  nhcl2 = 85.92,
+  ncl3 = 120.365,
   n = 14.0067,
   nh4oh = 18.04 + 17.008,
   nh42so4 = 18.04 * 2 + 96.0626,
@@ -785,3 +789,14 @@ bromatecoeffs <- data.frame(
     I = 1 # temp not in exponent
   )
 usethis::use_data(bromatecoeffs, overwrite = TRUE)
+
+# Data frame of Cl2 decay coefficients
+
+cl2coeffs <- tibble(
+  treatment = c('chlorine_raw', 'chlorine_coag', 'chloramine'),
+  a = c(-0.8147, -0.8404, -0.99),
+  b = c(-2.2808, -0.404, -0.015),
+  c = c(-1.2971, -0.9108, NA)
+)
+
+usethis::use_data(cl2coeffs, overwrite = TRUE)