+++ NEWS.md
@@ -1,6 +1,19 @@
-# tidywater 0.6.0
+# development version
+
+## New features
+* chlorine and chloramine decay
+
+# tidywater 0.6.2
+
+* CRAN resubmission.
+* Minor changes to DESCRIPTION and examples using `plan`
+
+# tidywater 0.6.1
 
 * Initial CRAN submission.
+* Fix R CMD check notes
+
+# tidywater 0.6.0
 
 ## New features
 * biofilter_toc updates the bdoc water slot